chore(features): move tests back to `.tests`
