+++ features/steps/.test/assets/deployment/Chart.yaml
@@ -1,0 +1,3 @@
+apiVersion: v1
+name: dummies
+version: 0.0.0
