+++ features/steps/.test/assets/directory/file1
@@ -1,0 +1,2 @@
+Foo
+Bar
