+++ features/steps/.test/assets/directory/file2
@@ -1,0 +1,2 @@
+Baz
+Qux
