+++ features/steps/.test/hooks.test.js
@@ -4,7 +4,8 @@ const { directory } = require('@toa.io/filesystem')
 const mock = require('@toa.io/mock')
 
 jest.mock('@cucumber/cucumber', () => mock.gherkin)
-require('./hooks')
+
+require('../hooks')
 
 const gherkin = mock.gherkin
 
@@ -13,10 +14,6 @@ let context
 
 beforeEach(async () => {
   context = { cwd: await directory.temp() }
-})
-
-afterEach(async () => {
-  if (context.cwd !== undefined) await directory.remove(context.cwd)
 })
 
 describe('Before', () => {
